ingestion/requirements.txt
requests
streamlit
pandas
python-dotenv
schedule
plotly
